reorganized code and cleaned metrics
